refactor: share stake card renderer via hook
